Second half of merge conflicts; added activity-handler to break out logic
