+++ demo/d2l-evaluation-hub/page-handler.js
@@ -32,86 +32,10 @@ function createPageEndpoint(activities, sorts, pageNumber, filtersHref, sortsHre
 		const sortState = decodeSortState(serializedSortState);
 
 		const sortedActivities = applySorts(activities, sorts, sortState);
-		const pagedActivities = chunk(sortedActivities.map(a => formatActivity(a)), 3);
+		const pagedActivities = chunk(sortedActivities, 3);
 
 		return formatPage(pagedActivities[pageNumber], filtersHref, sortsHref, addSortToHref(nextPageHref, sortState));
 	};
-}
-
-function formatActivity(activity) {
-	const formattedActivity = {
-		'class': [
-			activity.klass,
-			'activity'
-		],
-		'rel': [
-			'https://activities.api.brightspace.com/rels/user-activity-usage'
-		],
-		'links': [
-			{
-				'rel': [
-					'https://api.brightspace.com/rels/user'
-				],
-				'href': activity.userHref
-			},
-			{
-				'rel': [
-					'https://api.brightspace.com/rels/organization'
-				],
-				'href': activity.courseHref
-			},
-			{
-				'rel': [
-					activity.activityRel
-				],
-				'href': activity.activityHref
-			}
-		],
-		'entities': [
-			{
-				'class': [
-					'relative-uri'
-				],
-				'rel': [
-					'item',
-				],
-				'properties': {
-					'path': '/this/is/not/used'
-				}
-			},
-			{
-				'class': [
-					'date',
-					'localized-formatted-date'
-				],
-				'rel': [
-					'https://api.brightspace.com/rels/date'
-				],
-				'properties': {
-					'date': '2019-03-13T15:16:10.793Z',
-					'text': activity.localizedFormattedDate
-				}
-			}
-		]
-	};
-
-	if (activity.isDraft) {
-		const draftEntity = {
-			'class': [
-				'evaluation'
-			],
-			'rel': [
-				'https://api.brightspace.com/rels/evaluation'
-			],
-			'properties': {
-				'state': 'Draft'
-			}
-		};
-
-		formattedActivity.entities.push(draftEntity);
-	}
-
-	return formattedActivity;
 }
 
 function formatPage(entities, filterLocation, sortsLocation, nextLocation) {
